# billing-app/openapi.yaml
openapi: 3.1.0
info:
  title: Billing App API
  version: 0.1.0
  description: |
    Minimal API for billing setup and product management.
    The /setup endpoint supports both JSON (application/json) and HTML form submissions (application/x-www-form-urlencoded).
    Form submissions receive a 303 redirect on success/conflict; JSON clients receive JSON bodies and status codes.
servers:
  - url: http://localhost:8080
paths:
  /setup:
    head:
      summary: Quick status (headers only) if initial setup completed
      responses:
        '200':
          description: Returns X-Setup-Configured header (true|false)
          headers:
            X-Setup-Configured:
              description: Indicates whether company setup is completed
              schema: { type: string, enum: ["true", "false"] }
    get:
      summary: Check if setup already completed
      responses:
        '200':
          description: Setup status
          content:
            application/json:
              schema:
                type: object
                properties:
                  configured:
                    type: boolean
    post:
      summary: Perform initial company setup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetupRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/SetupRequest'
      responses:
        '201':
          description: Created company settings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompanySettings'
        '400': { description: Validation error }
        '409': { description: Already configured }
  /products:
    get:
      summary: List products
      responses:
        '200':
          description: Product list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
    post:
      summary: Create product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductInput'
      responses:
        '201': { description: Product created }
components:
  schemas:
    SetupRequest:
      type: object
      required: [company, address1, postal_code, city, country, siret]
      properties:
        company: { type: string }
        address1: { type: string }
        address2: { type: string }
        postal_code: { type: string }
        city: { type: string }
        country: { type: string }
        siret: { type: string }
        vat_enabled: { type: boolean }
        vat_rate: { type: number, format: float }
    ProductInput:
      type: object
      required: [name, price_ht]
      properties:
        name: { type: string }
        description: { type: string }
        price_ht: { type: number, format: float }
    Product:
      allOf:
        - $ref: '#/components/schemas/ProductInput'
        - type: object
          properties:
            id: { type: integer }
    CompanySettings:
      type: object
      properties:
        id: { type: integer }
        raison_sociale: { type: string }
        nom_commercial: { type: string }
        siren: { type: string }
        siret: { type: string }
        code_naf: { type: string }
        tva: { type: number }
        redevable_tva: { type: boolean }